+++ src/termux_tasker/ui/screens/bundled_runner.py
@@ -59,7 +59,7 @@ class BundledRunnerScreen(MenuScreen):
             is_installed = meta.general.id in installed_runners
             label = meta.general.name
             if is_installed:
-                label += " \\[Installed]"
+                label += " [$text-success]\\[Installed][/$text-success]"
             items.append(
                 ButtonConfig(
                     f"install_{meta.general.id}",